requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.5
pydantic==2.7.4
pydantic-settings==2.3.4
celery==5.4.0
redis==5.0.4
streamlit==1.35.0
openai==1.31.2
httpx<0.28.0
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.4
mammoth==1.11.0
